+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# v0.29.0
+
+- Rename `fetch_groups` to `fetch_all_groups`
+- Add `--json` flag to write logs to a `.jsonl` file
+
 # v0.28.0
 
 - Switch from using `fzf` to `PickerTui` everywhere